requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
influxdb-client==1.38.0

# Data Processing & Analysis
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
ta-lib==0.4.28
yfinance==0.2.28

# Machine Learning
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2
optuna==3.4.0

# Real-time Communication
websockets==12.0
aiohttp==3.9.1
kafka-python==2.0.2
celery==5.3.4

# Financial Data & APIs
ccxt==4.1.64
python-binance==1.0.19
alpha-vantage==2.3.1
finnhub-python==2.4.19

# Technical Analysis
TA-Lib==0.4.28
pandas-ta==0.3.14b0
stockstats==0.6.2

# Visualization & Charts
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Security & Authentication
cryptography==41.0.8
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Cloud & Infrastructure
boto3==1.34.0
docker==6.1.3

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
